Feature/143770291/set up client environment

+++ client/components/document/AllDocument.jsx
@@ -31,6 +31,7 @@ class AllDocument extends React.Component {
       document: {
         content: props.document.content || '',
         title: props.document.title || '',
+        access: props.document.access || '',
         id: props.document.id
       },
     };
@@ -44,9 +45,25 @@ class AllDocument extends React.Component {
    *
    * @return {void}
    */
+   
   deleteDocument() {
-    Materialize.toast('Document Deleted', 4000);
-    this.props.deleteDocument(this.props.document.id);
+     swal({
+      title: "Are you sure?", 
+      text: "Are you sure that you want to delete this document?", 
+      type: "error",
+      showCancelButton: true,
+      closeOnConfirm: true,
+      confirmButtonText: "Delete it!",
+      confirmButtonColor:  "#ad1457"
+    }, (isConfirm) => {
+      if(isConfirm) {
+         this.props.deleteDocument(this.props.document.id);
+         swal('Deleted', 'Document was deleted', 'success');
+         Materialize.toast('Document Deleted', 1000);
+      } else {
+        swal('Canceled', 'OPERATION CANCELED', 'error');
+      }
+    });
   }
 
   /**
@@ -86,7 +103,7 @@ class AllDocument extends React.Component {
    */
   render() {
     const { document } = this.props;
-    const { title, content } = this.state.document;
+    const { title, content, access } = this.state.document;
     const singleDocUrl = `document/${document.id}`;
     return (
       <div>
@@ -100,25 +117,11 @@ class AllDocument extends React.Component {
           <div className="card-action">
             {hasDocumentPermission(document.ownerId) ? <ul>
               <li>
-                <Modal
-                  id="mod"
-                  className="teal-text"
-                  trigger={
-                    <a className="btn-floating grey lighten-5 right">
+                    <a id="userDocDelete"
+                     onClick={this.deleteDocument}
+                     className="btn-floating grey lighten-5 right"
+                     >
                       <i className="material-icons red-text">delete</i></a>
-                       }
-                >
-                  <span className="delHeader"> Are You Sure You Want To Delete This Document ? </span>
-                  <div>
-                  <button
-                    onClick={this.deleteDocument}
-                    className="btn btn2 pink darken-4 white-text modal-action modal-close"
-                  >Yes</button>
-                  <button
-                    className="btn btn2 pink darken-4 white-text modal-action modal-close"
-                  >No</button>
-                  </div>
-                </Modal>
                 <Modal
                   header='Update Document'
                   className="teal-text"
@@ -130,7 +133,8 @@ class AllDocument extends React.Component {
                 >
                   <div className="row">
                     <form>
-                      <div className="input-field col s10">
+                     <label className="red-text">Title</label>
+                      <div className="input-field col s12">
                         <textarea
                           className="materialize-textarea"
                           name="title"
@@ -142,7 +146,8 @@ class AllDocument extends React.Component {
                   </div>
                   <div className="row">
                     <form>
-                      <div className="input-field col s10">
+                    <label className="red-text">Content</label>
+                      <div className="input-field col s12">
                         <textarea
                           className="materialize-textarea"
                           name="content"
@@ -152,9 +157,22 @@ class AllDocument extends React.Component {
                       </div>
                     </form>
                   </div>
+                   <label className="red-text">Access</label>
+                  <div className="row">
+                    <form>
+                      <div className="input-field col s12">
+                        <textarea
+                          className="materialize-textarea"
+                          name="access"
+                          value={access}
+                          onChange={e => this.updateDocumentState(e)}
+                        />
+                      </div>
+                    </form>
+                  </div>
                   <button
                     onClick={() => this.updateDocument()}
-                    className="btn pink darken-4 white-text modal-action modal-close"
+                    className="update-btn btn pink darken-4 white-text modal-action modal-close"
                   >Update</button>
                 </Modal>
               </li>
